feat(cache): refactor install azurite script
