+++ apps/cache-manager/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "build": "NODE_ENV=production webpack",
     "server": "ts-node -T ./src/server.ts",
-    "setup:local": "ts-node -T ./src/installAzurite.ts",
+    "azurite:install": "ts-node -T ./src/scripts/install-azurite.ts",
     "azurite:start": "azurite --location ./azurite --debug ./azurite/debug.log",
     "test": "node ../../scripts/test.js --coverage --watchAll=false",
     "test:watch": "node ../../scripts/test.js --watch"
